+++ crates/core/tests/java/deconstruction_list_wrap/case_expr.java
@@ -1,0 +1,9 @@
+class Deconstruction {
+    int f(Object o) {
+        int r = switch (o) {
+            case Point(int alpha, int beta) -> handle();
+            default -> 0;
+        };
+        return r;
+    }
+}