+++ crates/core/tests/java/deconstruction_list_wrap/case_group_wrap_if_long.out.java
@@ -1,0 +1,15 @@
+class Deconstruction {
+    void f(Object o) {
+        switch (o) {
+            case Point(
+                       int alphaComponent,
+                       int betaComponent,
+                       int gammaComponent
+            ):
+                handle();
+                break;
+            default:
+                break;
+        }
+    }
+}