+++ crates/core/tests/java/space_before_deconstruction_list/label.java
@@ -1,0 +1,8 @@
+class Deconstruction {
+    void f(Object o) {
+        switch (o) {
+            case Point(int alpha, int beta) -> handle();
+            default -> {}
+        }
+    }
+}